MicrophoneVolumeViewを破棄する時にクラッシュしていたので修正しました
